Get rid of version 1 stuff
