+++ lib/beautifier.js
@@ -4,9 +4,9 @@ const md = require('markdown-it')();
 
 const sharedStart = (array) => {
   const A = array.concat().sort();
-  const a1 = A[0], a2= A[A.length-1], L= a1.length;
+  const a1 = A[0], a2 = A[A.length - 1], L = a1.length;
   let i = 0;
-  while (i<L && a1.charAt(i)=== a2.charAt(i)) i++;
+  while (i < L && a1.charAt(i) === a2.charAt(i)) i++;
   return a1.substring(0, i);
 };
 
@@ -52,21 +52,92 @@ const generateExample = schema => {
 const beautifyMessage = (message) => {
   if (message.payload) message.payload = resolveAllOf(message.payload);
   if (message.headers) message.headers = resolveAllOf(message.headers);
-  message.summaryAsHTML = md.render(message.summary || '');
   message.descriptionAsHTML = md.render(message.description || '');
 
   if (message.headers) {
-    beautifySchema(message.headers);
-    message.generatedHeadersExample = stringify(generateExample(message.headers));
+    _.each(message.headers, beautifySchema);
+
+    message.formattedHeadersExamples = [];
+    if (message.examples) {
+      message.examples.forEach((example) => {
+        if (example.headers) {
+          message.hasHeadersExample = true;
+          message.formattedHeadersExamples.push(stringify(example.headers));
+        }
+      });
+    }
+
+    if (!message.formattedHeadersExamples.length) {
+      message.hasHeadersExample = true;
+      message.generatedHeadersExample = {};
+      _.each(message.headers, (header, headerName) => {
+        message.generatedHeadersExample[headerName] = generateExample(header);
+      });
+      message.generatedHeadersExample = stringify(message.generatedHeadersExample);
+    }
   }
+
   if (message.payload) {
     beautifySchema(message.payload);
-    if (message.payload.example) {
-      message.formattedExample = stringify(message.payload.example);
-    } else {
+
+    message.formattedPayloadExamples = [];
+    if (message.examples) {
+      message.examples.forEach((example) => {
+        if (example.payload) {
+          message.hasPayloadExample = true;
+          message.formattedPayloadExamples.push(stringify(example.payload));
+        }
+      });
+    }
+
+    if (!message.formattedPayloadExamples.length) {
+      message.hasPayloadExample = true;
       message.generatedPayloadExample = stringify(generateExample(message.payload));
     }
   }
+
+  if (message.oneOf) {
+    message.formattedHeadersExamples = [];
+    message.formattedPayloadExamples = [];
+
+    message.oneOf.forEach(msg => {
+      if (msg.headers) _.each(msg.headers, beautifySchema);
+      if (msg.payload) beautifySchema(msg.payload);
+
+      if (msg.examples) {
+        msg.examples.forEach((example) => {
+          if (example.headers) {
+            message.hasHeadersExample = true;
+            message.formattedHeadersExamples.push(stringify(example.headers));
+          }
+
+          if (example.payload) {
+            message.hasPayloadExample = true;
+            message.formattedPayloadExamples.push(stringify(example.payload));
+          }
+        });
+      }
+    });
+
+    if (!message.formattedHeadersExamples.length && message.oneOf[0].headers) {
+      message.hasHeadersExample = true;
+      message.generatedHeadersExample = {};
+      _.each(message.oneOf[0].headers, (header, headerName) => {
+        message.generatedHeadersExample[headerName] = generateExample(header);
+      });
+      message.generatedHeadersExample = stringify(message.generatedHeadersExample);
+    }
+
+    if (!message.formattedPayloadExamples.length) {
+      const payload = message.oneOf[0].payload;
+      if (payload) {
+        message.hasPayloadExample = true;
+        message.generatedPayloadExample = stringify(generateExample(payload));
+      }
+    }
+  }
+
+  message.uid = Buffer.from(JSON.stringify(message)).toString('base64');
 
   return message;
 };
@@ -83,95 +154,74 @@ const beautifySchema = (schema) => {
 };
 
 module.exports = (asyncapi) => {
-  asyncapi.baseTopic = asyncapi.baseTopic || '';
   asyncapi.info = asyncapi.info || {};
   asyncapi.info.descriptionAsHTML = md.render(asyncapi.info.description || '');
 
   if (asyncapi.servers) {
     _.each(asyncapi.servers, server => {
       server.descriptionAsHTML = md.render(server.description || '');
+
+      if (server.security) {
+        server._security = [];
+
+        _.each(server.security, security => {
+          const name = Object.keys(security)[0];
+          if (!asyncapi.components || !asyncapi.components.securitySchemes || !asyncapi.components.securitySchemes[name]) {
+            throw new Error(`Security definition "${name}" is not in included in #/components/securitySchemes.`);
+          }
+
+          asyncapi.components.securitySchemes[name].descriptionAsHTML = md.render(asyncapi.components.securitySchemes[name].description || '');
+          server._security.push(asyncapi.components.securitySchemes[name]);
+        });
+      }
+
       _.each(server.variables, variable => {
         variable.descriptionAsHTML = md.render(variable.description || '');
       });
     });
   }
 
-  if (asyncapi.security) {
-    asyncapi._security = [];
+  if (asyncapi.channels) {
+    asyncapi.subscribeChannels = {};
+    asyncapi.publishChannels = {};
+  }
 
-    _.each(asyncapi.security, security => {
-      const name = Object.keys(security)[0];
-      if (!asyncapi.components || !asyncapi.components.securitySchemes || !asyncapi.components.securitySchemes[name]) {
-        throw new Error(`Security definition "${name}" is not in included in #/components/securitySchemes.`);
+  _.each(asyncapi.channels, (channel, channelName) => {
+    if (channel.subscribe) {
+      if (channel.subscribe.message) {
+        channel.subscribe.descriptionAsHTML = md.render(channel.subscribe.description || '');
+        beautifyMessage(channel.subscribe.message);
+      } else if (channel.subscribe.message && channel.subscribe.message.oneOf) {
+        _.each(channel.subscribe.message.oneOf, beautifyMessage);
       }
 
-      asyncapi.components.securitySchemes[name].descriptionAsHTML = md.render(asyncapi.components.securitySchemes[name].description || '');
-      asyncapi._security.push(asyncapi.components.securitySchemes[name]);
-    });
-  }
-
-  if(asyncapi.topics){
-    asyncapi.subscribeTopics = {};
-    asyncapi.publishTopics = {};
-  }
-
-  _.each(asyncapi.topics, (topic, topicName) => {
-    const separator = asyncapi['x-topic-separator'] || '.';
-    const baseTopic = asyncapi.baseTopic.trim();
-
-    const newTopicName = baseTopic.length ? `${baseTopic}${separator}${topicName}` : topicName;
-    if (newTopicName !== topicName) {
-      asyncapi.topics[newTopicName] = topic;
-      delete asyncapi.topics[topicName];
+      asyncapi.subscribeChannels[channelName] = channel;
     }
 
-    if (topic.publish) {
-      beautifyMessage(topic.publish);
-      if (topic.publish.oneOf) _.each(topic.publish.oneOf, beautifyMessage);
-      asyncapi.publishTopics[newTopicName] = topic;
+    if (channel.publish) {
+      if (channel.publish.message) {
+        channel.publish.descriptionAsHTML = md.render(channel.publish.description || '');
+        beautifyMessage(channel.publish.message);
+      } else if (channel.publish.message && channel.publish.message.oneOf) {
+        _.each(channel.publish.message.oneOf, beautifyMessage);
+      }
 
+      asyncapi.publishChannels[channelName] = channel;
     }
 
-    if (topic.subscribe) {
-      beautifyMessage(topic.subscribe);
-      if (topic.subscribe.oneOf) _.each(topic.subscribe.oneOf, beautifyMessage);
-      asyncapi.subscribeTopics[newTopicName] = topic;
-    }
-
-    if (topic.parameters) {
-      topic.parameters.forEach((param) => {
+    if (channel.parameters) {
+      channel.parameters.forEach((param) => {
         param.descriptionAsHTML = md.render(param.description || '');
       });
     }
   });
 
   if (!asyncapi.components) asyncapi.components = {};
-  if (!asyncapi.components.messages) {
-    asyncapi.__noMessages = true;
-    asyncapi.components.messages = {};
-  }
-  if (!asyncapi.components.schemas) {
-    asyncapi.__noSchemas = true;
-    asyncapi.components.schemas = {};
-  }
 
-  _.each(asyncapi.components.messages, beautifyMessage);
-
-  _.each(asyncapi.components.schemas, (schema, schemaName) => {
-    schema = resolveAllOf(schema);
-    if (schema.example) {
-      schema.formattedExample = stringify(schema.example);
-    } else {
-      schema.generatedExample = stringify(generateExample(schema));
-    }
-    asyncapi.components.schemas[schemaName] = schema;
-    beautifySchema(schema);
-  });
-
-  if (asyncapi.topics) {
-    const commonPrefix = sharedStart(Object.keys(asyncapi.topics));
-    const levels = commonPrefix.split('.').length - 1;
-    asyncapi.__commonPrefix = commonPrefix.split('.').slice(0, levels).join('.');
+  if (asyncapi.channels) {
+    const commonPrefix = sharedStart(Object.keys(asyncapi.channels));
+    const levels = commonPrefix.split('/').length - 1;
+    asyncapi.__commonPrefix = commonPrefix.split('/').slice(0, levels).join('/');
   }
 
   return asyncapi;
