+++ lib/generator.js
@@ -7,7 +7,6 @@ const randomName = require('project-name-generator');
 const xfs = require('fs.extra');
 const { parse } = require('asyncapi-parser');
 const beautifier = require('./beautifier');
-const beautifier2 = require('./beautifier2');
 const registerPartial = require('./register-partial');
 const minimatch = require("minimatch");
 
@@ -282,7 +281,7 @@ const registerPartials = config => new Promise((resolve, reject) => {
 const parseAndApplyDefaults = config => new Promise((resolve, reject) => {
   try {
     const parsed = parse(config.asyncapi);
-    config.asyncapi = beautifier2(parsed);
+    config.asyncapi = beautifier(parsed);
     const random_name = randomName().dashed;
 
     if (!config.template) {
